+++ tests/test_scripts.py
@@ -2,16 +2,17 @@ import pytest
 import pathlib
 import hashlib
 
+from ibex_imaging_knowledge_base_utilities.bib2md import bibfile2md
+from ibex_imaging_knowledge_base_utilities.zenodo_json_2_thewho_md import zenodo_creators_to_md
+from ibex_imaging_knowledge_base_utilities.datadict_glossary_2_contrib_md import dict_glossary_to_md
 from ibex_imaging_knowledge_base_utilities.reagent_resources_csv_2_md_url import (
     csv_to_md_with_url,
 )
-from ibex_imaging_knowledge_base_utilities.fluorescent_probes_csv_2_md import (
-    fluorescent_probe_csv_to_md,
-)
-
-from ibex_imaging_knowledge_base_utilities.bib2md import bibfile2md
 from ibex_imaging_knowledge_base_utilities.update_index_md_stats import (
     update_index_stats,
+)
+from ibex_imaging_knowledge_base_utilities.fluorescent_probes_csv_2_md import (
+    fluorescent_probe_csv_to_md,
 )
 from ibex_imaging_knowledge_base_utilities.validate_zenodo_json import (
     validate_zenodo_json,
@@ -149,4 +150,22 @@ class ZenodoJson2Contrib(BaseTest):
     def test_zenodo_creators_to_md(self, input_md_file_name, zenodo_json_file_name, result_md5hash, tmp_path):
         output_dir = tmp_path
         zenodo_creators_to_md(self.data / input_md_file_name, self.data_path / zenodo_json_file_name, output_dir)
+        assert self.files_md5([output_dir / pathlib.Path(input_md_file_name).stem]) == result_md5hash
+
+class TestDictGlossary2Contrib(BaseTest):
+    @pytest.mark.parametrize(
+        "input_md_file_name, dict_csv_file_name, glossary_csv_file_name, result_md5hash",
+        [("contrib.md.in ", "reagent_data_dict.csv", "reagent_glossary.csv", "c914b681cf98941d9e8f74a4fe4bd892")],
+    )
+    def test_dict_glossary_to_md(
+            self, input_md_file_name, dict_csv_file_name, glossary_csv_file_name, result_md5hash, tmp_path
+    ):
+        # Write the output using the tmp_path fixture
+        output_dir = tmp_path
+        dict_glossary_to_md(
+            self.data_path / input_md_file_name,
+            self.data_path / dict_csv_file_name,
+            self.data_path / glossary_csv_file_name,
+            output_dir,
+        )
         assert self.files_md5([output_dir / pathlib.Path(input_md_file_name).stem]) == result_md5hash